% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_menstruation_cycle.R
\name{get_menstruation_cycle}
\alias{get_menstruation_cycle}
\title{Get Menstruation Cycle Data from Apple Health Database}
\usage{
get_menstruation_cycle(
  health_db,
  type = "HKCategoryTypeIdentifierMenstrualFlow",
  start_date = NULL,
  end_date = NULL
)
}
\arguments{
\item{health_db}{A dataframe or the path to an RDS file containing the health data.}

\item{type}{The type of menstruation data to extract, with default being "HKCategoryTypeIdentifierMenstrualFlow".}

\item{start_date}{The start of the time frame for which to extract data.}

\item{end_date}{The end of the time frame for which to extract data.}
}
\value{
A dataframe with a sequence of dates and a flag for menstruation days.
}
\description{
This function identifies menstruation periods from the Apple Health Database, either passed as a dataframe
or loaded from an RDS file, and returns a dataframe with a complete sequence of dates and a flag
indicating menstruation days within the range where menstruation data exists.
}
\note{
This function assumes the health data is in UTC time zone and that menstruation data includes 'startDate' and 'endDate'.
}
\examples{
\dontrun{
  library(tidyverse)
  health_data <- read_rds("path/to/your/health_data.rds")
  menstruation_cycle_data <- get_menstruation_cycle(health_db = health_data)
}
}
\references{
Use the Apple Health documentation for reference to the data structure.
}
