% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vo2max.R
\name{get_vo2max}
\alias{get_vo2max}
\title{Get VO2Max Data from Apple Health Database}
\usage{
get_vo2max(health_db, start_date = NULL, end_date = NULL)
}
\arguments{
\item{health_db}{A dataframe or the path to an RDS file containing the health data.}

\item{start_date}{The start of the time frame for which to extract VO2Max data.}

\item{end_date}{The end of the time frame for which to extract VO2Max data.}
}
\value{
A dataframe containing the filtered VO2Max data.
}
\description{
This function extracts VO2Max data for a specified time frame from the Apple Health Database,
either passed as a dataframe or loaded from an RDS file.
}
\note{
This function assumes the health data is in the local time zone of the data source.
}
\examples{
\dontrun{
  library(tidyverse)
  health_data <- read_rds("path/to/your/health_data.rds")
  vo2max_data <- get_vo2max(health_db = health_data)
}
}
\references{
Use the Apple Health documentation for reference to the data structure.
}
