% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_workouts.R
\name{import_workouts}
\alias{import_workouts}
\title{Import Workout Data from Apple Health Export}
\usage{
import_workouts(filename, datasource = "apple")
}
\arguments{
\item{filename}{The path to the zip file containing Apple Health Export data.}

\item{datasource}{The source of the health data, default is set to 'apple'.}
}
\value{
A named list of tibbles where each tibble contains columns for
latitude (lat), longitude (lon), elevation (ele), time, speed, course,
horizontal accuracy (hAcc), and vertical accuracy (vAcc). The names of the
list elements correspond to the names of the workout routes.
}
\description{
This function processes Apple Health Export data from a specified zip file.
It extracts GPS data from GPX files within the zip archive, parses the GPX
data, and compiles it into a named list of tibbles where each tibble
represents a workout session.
}
\examples{
# Assuming 'workouts.zip' is the Apple Health Export zip file in the working directory:
workout_data <- parse_workouts(filename = "workouts.zip")
}
