% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_healthdata.R
\name{parse_healthdata}
\alias{parse_healthdata}
\title{Parse Apple Health Data in Zip format}
\usage{
parse_healthdata(filename, type = "apple")
}
\arguments{
\item{filename}{filename of the raw data. For Apple Health Data, this should be a zip file}

\item{type}{type of health data. Currently supports 'Apple'}
}
\value{
data frame with health data
}
\description{
Parse Apple Health Data in Zip format
}
\note{
note
}
\examples{
options(tidyverse.quiet = TRUE)
library(tidyverse)
library(health2R)
further examples how to use here 
}
\references{
reference data structure
}
