% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_function.R
\name{test_function}
\alias{test_function}
\title{xxx}
\usage{
test_function(xx, xxx)
}
\arguments{
\item{xx}{xxxx}

\item{xxxx}{yyyyy}
}
\value{
zzzzz
}
\description{
xxx
}
\note{
note
}
\examples{
options(tidyverse.quiet = TRUE)
library(tidyverse)
library(health2R)
}
\references{
lalal
}
